+++ mesos/datastore.go
@@ -119,6 +119,10 @@ func (e *Scheduler) connectPort(host string, port uint32) bool {
 // set mysql parameter of the mesos task
 func (e *Scheduler) setMySQL(cmd *mesosutil.Command) {
 	cmd.ContainerImage = e.Config.ImageMySQL
+	// Enable TLS for Mariadb
+	if e.Config.DSMySQLSSL {
+		cmd.Command = "/usr/bin/mariadb --ssl-ca=/var/lib/mysql/root-ca.pem --ssl-cert=/var/lib/mysql/server-cert.pem --ssl-key=/var/lib/mysql/server-key.pem"
+	}
 	cmd.Environment.Variables = []mesosproto.Environment_Variable{
 		{
 			Name:  "SERVICE_NAME",